[brain tests] Burst six from the main file
